added logging
Fixes #10. With this PR it is now supported to send STDOUT logs to Elastic or Splunk. LogSpec struct has been added to the Capp CRD. In the case of elastic type, It is required to enter host, sslverify, username, and password secret name. An index is also optional. In the case of Splunk type, it is required to enter host, sslverify, and hec token. An index is also optional. This PR also includes the addition of capp-Logging tests.

Signed-off-by: liadlauber <liad6500@gmail.com>